Move unit metatable stuff into lua_unit file
